requirements.txt
opencv-python == 4.6.0.66
PyQt5-stubs==5.15.6.0
pandas ==1.5.1
Pillow == 9.3.0
scipy==1.8.0
numpy==1.22.3